Improves ambiguous file handling
Adds enhanced logging for ambiguous file routing and creates json
metadata to describe why files are being routed to the ambiguous
directory, what the best candidate was, and what other candidates
were considered.

Adds a warning to file upload actions in github, so that the workflow
doesn't fail because of missing coverage or junit xml files.

Fixes a race condition when acquiring the web port.