Fixes file check and queue conditions
Corrects a condition to verify the existence of the target
movie file before enqueuing a work command. This prevents
the processing of non-existent files.

Also, corrects a condition when adding commands to the work
queue, addressing issues where valid commands were not
being added when the server is not stopping.