Corrects command queue logic (#116)
* Fixes file check and queue conditions

Corrects a condition to verify the existence of the target
movie file before enqueuing a work command. This prevents
the processing of non-existent files.

Also, corrects a condition when adding commands to the work
queue, addressing issues where valid commands were not
being added when the server is not stopping.

* Fixes command queue logic

Fixes an issue where commands were not added to the work queue
when the target was already in the queue, and the command was None.

Also adds a logger to catch queue work exceptions.